refactor: Move source/httpsource to source/type/httpsource
