+++ internal/pkg/service/stream/source/type/httpsource/config.go
