+++ internal/pkg/service/stream/source/type/httpsource/error.go
